fix prototype for skip, OK

+++ src/JSystem/JSUInputStream.hpp
@@ -6,7 +6,7 @@ public:
 	virtual ~JSUInputStream();
 	virtual void getAvailable() = 0;
 	void read(void *, s32);
-	virtual void skip(u8);
+	virtual void skip(u16);
 	u32 _000;
 	u32 _004;
 };
